refactor: move parse response logic to own module
